% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_bowtie.r
\name{bowtie_align}
\alias{bowtie_align}
\title{RACEseq BOWTIE alignment function}
\usage{
bowtie_align(input_data, replicon_ref, mismatch, RACE_adapter, out_name)
}
\arguments{
\item{input_data}{The input .fastq files to be used for the alignment.}

\item{replicon_ref}{The reference sequence to be used in the alignment.}

\item{mismatch}{The number of mismatches to be allowed during alignment.}

\item{RACE_adapter}{The RACE adapter nucleotide sequence to be trimmed from the reads using cutadapt.}

\item{out_name}{The output filename from the alignment.}
}
\description{
Function to perform the trimming of the RACE adapter using CUTADAPT and the alignment of the sequencing data from the .fastq files using the BOWTIE aligner.
}
\examples{
bowtie_align()
}
\seealso{
tmap_align
}
\keyword{alignment}
\keyword{bowtie}
\keyword{cutadapt}
