% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_out.r
\name{datafile_out}
\alias{datafile_out}
\title{RACEseqR dataframe function}
\usage{
datafile_out(str, end, replicon_ref, filename)
}
\arguments{
\item{str}{The start position of the binding region.}

\item{end}{The end position of the binding region.}

\item{replicon_ref}{The reference sequence used in the alignment.}

\item{filename}{The filename that will be used for the output data file.}
}
\description{
Function to generate and output the RACE seq dataframe of the specified start and end positions.
}
\examples{
out_csv()
}
\keyword{dataframe}
\keyword{output}
