% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del.r
\name{del_files}
\alias{del_files}
\title{File delete function}
\usage{
del_files(pattern)
}
\arguments{
\item{pattern}{The filename extension to be deleted.}
}
\description{
Function to delete files generated in the process.
}
\examples{
del_files()
}
\keyword{pattern}
