% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.r
\name{out_csv}
\alias{out_csv}
\title{RACEseqR dataframe function}
\usage{
out_csv(str, end)
}
\arguments{
\item{str}{The start position of the binding region.}

\item{end}{The end position of the binding region.}
}
\description{
Function to generate and output the RACE seq dataframe of the specified start and end positions.
}
\examples{
out_csv()
}
\keyword{binding_region}
