% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.r
\name{packages}
\alias{packages}
\title{Package download function}
\usage{
packages(x)
}
\arguments{
\item{x}{The name of the package you wish to download and install? Defaults NULL.}
}
\description{
Function to download, install and load the required libraries only when needed
}
\examples{
packages()
}
\keyword{cats}
