% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplot.r
\name{plot_out}
\alias{plot_out}
\title{RACEseqR plot function}
\usage{
plot_out(binding_region, filename)
}
\arguments{
\item{binding_region}{The region that will be plotted in the graph. Can be an R dataframe or a tab delim file in your working directory.}

\item{filename}{The filename that will be used for the output pdf file.}
}
\description{
Function to generate and output the RACE seq plot of the specified start and end positions.
}
\examples{
out_plot()
}
\keyword{binding_region}
