% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.r
\name{RACEseqR}
\alias{RACEseqR}
\title{RACEseqR main function}
\usage{
RACEseqR(str, end, mismatch, RACE_adapter, tmap_opt)
}
\arguments{
\item{str}{The start position of the binding region.}

\item{end}{The end position of the binding region.}

\item{mismatch}{The number of mismatches to be accepted during alignment.}

\item{RACE_adapter}{The RACE adapter to be trimmied using cutadapt .}

\item{tmap_opt}{The option to use the Tmap aligner. Defaults FALSE}
}
\description{
Function to generate the alignment from a RACE sequencing experiment
}
\examples{
RACEseq()
}
\keyword{binding_region}
