% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.r
\name{tmap_align}
\alias{tmap_align}
\title{RACEseqR TMAP alignment function}
\usage{
tmap_align(input_data, replicon_ref, mismatch, out_name)
}
\arguments{
\item{input_data}{The input .fastq files to be used for the alignment.}

\item{replicon_ref}{The reference sequence to be used in the alignment.}

\item{mismatch}{The number of mismatches to be allowed during alignment.}

\item{out_name}{The output filename from the alignment.}
}
\description{
Function to perform the alignment of the sequencing data from the .fastq files using the TMAP aligner.
}
\examples{
tmap_align()
}
\seealso{
bowtie_align

bowtie_trim_align

tmap_trim_align
}
\keyword{alignment}
\keyword{tmap}
