% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.r
\name{out_csv}
\alias{out_csv}
\title{RACEseq dataframe function}
\usage{
out_csv(binding_region, filename)
}
\arguments{
\item{binding_region}{The region that will be written in the csv file. Defaults NULL.}

\item{filename}{The filename of the dataset}
}
\description{
Function to generate and output the RACE seq dataframe of the specified start and end positions.
}
\examples{
out_csv()
}
\keyword{binding_region}
