% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_out.r, R/main_fun.r
\name{out_csv}
\alias{out_csv}
\alias{out_csv}
\title{RACEseqR dataframe function}
\usage{
out_csv(str, end)

out_csv(str, end)
}
\arguments{
\item{str}{The start position of the binding region.}

\item{end}{The end position of the binding region.}

\item{df_data}{The read count data generated from the alignment}

\item{replicon_ref}{The reference sequence used in the alignment.}

\item{str}{The start position of the binding region.}

\item{end}{The end position of the binding region.}
}
\description{
Function to generate and output the RACE seq dataframe of the specified start and end positions.

Function to generate and output the RACE seq dataframe of the specified start and end positions.
}
\examples{
out_csv()
out_csv()
}
\keyword{binding_region}
\keyword{dataframe}
\keyword{output}
