% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.r
\name{RACEseq}
\alias{RACEseq}
\title{RACEseq main function}
\usage{
RACEseq(input_data, replicon_ref, mismatch = 0, RACE_adapter = NULL, str,
  end, filename, tmap)
}
\arguments{
\item{input_data}{The input .fastq files to be used for the alignment.}

\item{replicon_ref}{The reference sequence to be used in the alignment.}

\item{mismatch}{The number of mismatches to be accepted during alignment.}

\item{RACE_adapter}{The RACE adapter to be trimmied using cutadapt .}

\item{str}{The start position of the binding region.}

\item{end}{The end position of the binding region.}

\item{filename}{The filename that will be used for the output pdf file.}

\item{tmap}{The option to use the Tmap aligner. Defaults NULL}
}
\description{
Function to generate the alignment from a RACE sequencing experiment
}
\examples{
RACEseq()
}
\keyword{binding_region}
