% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.r
\name{RACEseq}
\alias{RACEseq}
\title{RACEseqR main function}
\usage{
RACEseq(mismatch, RACE_adapter, tmap_opt)
}
\arguments{
\item{mismatch}{The number of mismatches to be accepted during alignment.}

\item{RACE_adapter}{The RACE adapter to be trimmied using cutadapt .}

\item{tmap_opt}{The option to use the Tmap aligner. Defaults FALSE}
}
\description{
Function to generate the alignment from a RACE sequencing experiment using the BOWTIE aligner
}
\details{
The The following UNIX command line programs and packages must be installed on your computer separately.

Package  Version

Cutadapt (optional)  >= 1.12

Bowtie 1.0.0

Samtools >=1.3.1

Bedtools >=2.17.0

Tmap (optional)  >=3.4.1
}
\examples{
RACEseq()
}
\keyword{binding_region}
