% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun_samefile.r
\name{RACEseq}
\alias{RACEseq}
\title{RACEseqR main function}
\usage{
RACEseq(input_data, replicon_ref, filename, mismatch = 0, RACE_adapter,
  str = NULL, end = NULL, tmap = NULL, plot = NULL)
}
\arguments{
\item{input_data}{The input .fastq files to be used for the alignment.}

\item{replicon_ref}{The reference sequence to be used in the alignment.}

\item{filename}{The filename that will be used for the output pdf file.}

\item{mismatch}{The number of mismatches to be accepted during alignment. DEFAULT = 0}

\item{RACE_adapter}{The RACE adapter to be trimmied using cutadapt.}

\item{str}{The start position of the binding region.}

\item{end}{The end position of the binding region.}

\item{tmap}{The option to use the Tmap aligner instead of the default BOWTIE aligner. DEFAULT = FALSE}

\item{plot}{The option to output the results in plot. The default output is an R dataframe object. DEFAULT = FALSE}
}
\description{
Function to generate the alignment from a RACE sequencing experiment
}
\examples{
RACEseq(filename = "foo", str = 9478, end = 9498, plot = T)
region_foo <- RACEseq(input_data = foo.fastq, replicon_ref = "foo.fasta", filename = "bar", RACE_adapter = "GGACACTGACATGGACTGAAGGAGTAGAAA", str = 9478, end = 9498)
}
\keyword{binding_region}
